Start on ObstacleCourse level, make it run in visual studio on windows

+++ src/very_general.h
@@ -5,13 +5,13 @@
 
 #define DEV 1
 // Whether to cycle through all levels quickly at beginning, to test their validity
-// #define TEST_ALL_LEVELS
+//#define TEST_ALL_LEVELS
 #define GAME_FPS 60
 
 #define WINDOW_WIDTH 840
 #define WINDOW_HEIGHT (630)
 
-typedef int8_t Coord; // Coord in world
+typedef int16_t Coord; // Coord in world
 
 typedef int32_t Int; // for general things like indexing arrays and stuff
